IOS-108 added FilterSheetModifier
